Update ClubCreationModal type name
