+++ components/ClubCreationModal/Common/Navigation/style.ts
@@ -1,4 +1,4 @@
-import { SelectType } from '@/type/components/ClubCreationModal'
+import { SelectType } from '@/type/components/ClubCreationPage'
 import styled from '@emotion/styled'
 
 export const Wrapper = styled.nav`
