+++ type/components/ClubCreationPage.ts
@@ -2,9 +2,22 @@ export interface SelectType {
   select: boolean
 }
 
-export type ClubImgUploadType = 'banner' | 'activity'
+export interface OptionProps extends SelectType {
+  url: string
+  position: string
+}
+
+export interface ErrorProps {
+  error?: boolean
+}
 
 export interface ImgUploadFormType {
   bannerImg: FileList
   activityImgs: FileList
 }
+
+export interface ContentFormType {
+  content: string
+}
+
+export type ClubImgUploadType = 'banner' | 'activity'
